Add Base Project Structure; Prototype setup.sh
Created the project's folder layout. This layout will be changed as
needed during the development of amiibrOS-arcade.

Created a proof-of-concept setup.sh. It lacks copying the amiibrOS
software and other file-structure-related specifics.

Added various README.md files to the project's subfolders. These
are unfinished.

Added a main README.md which is also unfinished.