Complete Boot Sequence Prototype
The setup.sh script has been finished. It flashes a base Raspbian
Lite image and modifies it so that upon boot, the Pi installs all
necessary packages and configurations for amiibrOS system software
to run. It then runs the prototype main.py.

Note, documentation is still out of sync with these changes.